Add support for decryption of binary asset files
tools/encrypt-asset.py can be used to encrypt an asset, and then
Glv.db will decrypt it in-place (actually, a direct buffer with
its data).

The key is hard-coded, hence this is more like obfuscation rather than
encryption.

Note that the current implementation appends zeros to the end of the
(decrypted) asset, which means that the original asset format must
support such trailing zeroes.