ColorTransferProcessor: initialize the output mat with the right type.
Fix a bug that the returned matrix was different of the given input mat.

The output mat type was defined at the moment the first pixel was
added/pushed to the output mat. In the case the mask was empty, no pixel
was added/pushed and thus the output mat type was never defined and
stayed with the default type (CvType.CV_8UC1) regardless the type of
the input mat.